utworzenie klasy EqulizerWithBands i przeniesienie odpowiedzialności za tworzenie części łańcucha audio do klasy EqSectionHandler
